+++ src/EntityGraphQL/Directives/IncludeDirectiveProcessor.cs
@@ -1,10 +1,13 @@
 using System;
+using System.ComponentModel;
 using EntityGraphQL.Compiler;
 
 namespace EntityGraphQL.Directives
 {
     public class IncludeDirectiveProcessor : DirectiveProcessor<IncludeArguments>
     {
+        public override string Name { get => "include"; }
+        public override string Description { get => "Directs the executor to include this field or fragment only when the `if` argument is true."; }
         public override bool ProcessesResult { get => false; }
         public override Type GetArgumentsType()
         {
@@ -21,6 +24,7 @@ namespace EntityGraphQL.Directives
 
     public class IncludeArguments
     {
+        [Description("Included when true.")]
         public bool @if { get; set; }
     }
 }