+++ src/EntityGraphQL/Directives/SkipDirectiveProcessor.cs
@@ -1,10 +1,13 @@
 using System;
+using System.ComponentModel;
 using EntityGraphQL.Compiler;
 
 namespace EntityGraphQL.Directives
 {
     public class SkipDirectiveProcessor : DirectiveProcessor<SkipArguments>
     {
+        public override string Name { get => "skip"; }
+        public override string Description { get => "Directs the executor to skip this field or fragment when the `if` argument is true."; }
         public override bool ProcessesResult { get => false; }
         public override Type GetArgumentsType()
         {
@@ -21,6 +24,7 @@ namespace EntityGraphQL.Directives
 
     public class SkipArguments
     {
+        [Description("Excluded when true.")]
         public bool @if { get; set; }
     }
 }