+++ src/EntityGraphQL/Schema/SchemaProvider.cs
@@ -58,7 +58,7 @@ namespace EntityGraphQL.Schema
             AddType<Models.TypeElement>("__Type", "Information about types").AddAllFields();
             AddType<Models.EnumValue>("__EnumValue", "Information about enums").AddAllFields();
             AddType<Models.InputValue>("__InputValue", "Arguments provided to Fields or Directives and the input fields of an InputObject are represented as Input Values which describe their type and optionally a default value.").AddAllFields();
-            AddType<Models.Directives>("__Directive", "Information about directives").AddAllFields();
+            AddType<Models.Directive>("__Directive", "Information about directives").AddAllFields();
             AddType<Models.Field>("__Field", "Information about fields").AddAllFields();
             AddType<Models.SubscriptionType>("Information about subscriptions").AddAllFields();
             AddType<Models.Schema>("__Schema", "A GraphQL Schema defines the capabilities of a GraphQL server. It exposes all available types and directives on the server, as well as the entry points for query, mutation, and subscription operations.").AddAllFields();
@@ -527,6 +527,7 @@ namespace EntityGraphQL.Schema
                     type = type.GetGenericArguments()[0];
                 }
             }
+
             if (customTypeMappings.ContainsKey(type))
                 return customTypeMappings[type];
 
@@ -667,6 +668,10 @@ namespace EntityGraphQL.Schema
             if (directives.ContainsKey(name))
                 return directives[name];
             throw new EntityGraphQLCompilerException($"Directive {name} not defined in schema");
+        }
+        public IEnumerable<IDirectiveProcessor> GetDirectives()
+        {
+            return directives.Values.ToList();
         }
         public void AddDirective(string name, IDirectiveProcessor directive)
         {